fix: infinite loop
